fixed overflow bug with chat by applying overflow to grandparent container
