+++ raynian-fe/package.json
@@ -24,6 +24,7 @@
     "react-tooltip": "^5.21.4",
     "redux": "^4.2.1",
     "socket.io-client": "^4.7.1",
+    "tailwind-scrollbar": "^3.0.5",
     "tinycolor2": "^1.6.0",
     "use-debounce": "^9.0.4",
     "validator": "^13.9.0",
