+++ raynian-fe/src/components/chat/Chat.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+import ChatDisplay from "./ChatDisplay";
+import ChatInput from "./ChatInput";
+
+const Chat = () => {
+  return (
+    <div>
+      <ChatDisplay />
+      <ChatInput />
+    </div>
+  );
+};
+
+export default Chat;
