+++ raynian-fe/src/components/chat/ChatInput.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const ChatInput = () => {
+  return <div>ChatInput</div>;
+};
+
+export default ChatInput;
